services/worker/requirements.txt
pyyaml
redis

